+++ src/net/sf/mzmine/userinterface/dialogs/experimentalparametersetupdialog/ExperimentalParametersSetupDialog.java
@@ -147,7 +147,7 @@ public class ExperimentalParametersSetupDialog extends JDialog implements Action
 			SimpleParameter parameter = null;
 
 			if (radiobuttonNumerical.isSelected()) {
-				Parameter.ParameterType paramType = Parameter.ParameterType.DOUBLE;
+				Parameter.ParameterType paramType = Parameter.ParameterType.FLOAT;
 				Double minValue = Double.NEGATIVE_INFINITY;
 				if (fieldNumericalMinValue.getValue()!=null)
 					minValue = ((Number)fieldNumericalMinValue.getValue()).doubleValue();
@@ -264,7 +264,7 @@ public class ExperimentalParametersSetupDialog extends JDialog implements Action
 		for (int columnIndex=0; columnIndex<parameterValues.keySet().size(); columnIndex++) {
 			Parameter parameter = tablemodelParameterValues.getParameter(columnIndex+1);
 			
-			if (parameter.getType()==Parameter.ParameterType.DOUBLE) {
+			if (parameter.getType()==Parameter.ParameterType.FLOAT) {
 				Double minValue = null;
 				Double maxValue = null;
 				if (parameter.getMinimumValue()!=null)
@@ -318,7 +318,7 @@ public class ExperimentalParametersSetupDialog extends JDialog implements Action
 				RawDataFile file = dataFiles[dataFileIndex];
 				
 				Object value = tablemodelParameterValues.getValueAt(dataFileIndex, columnIndex+1);
-				if (parameter.getType()==Parameter.ParameterType.DOUBLE) {
+				if (parameter.getType()==Parameter.ParameterType.FLOAT) {
 					Double doubleValue=null;
 					if (value instanceof Double)
 						doubleValue = (Double)value;
